# api/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Self-hosted media server for movies, TV shows, and music",
        "title": "ViewRA Media Server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "ViewRA Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/music/albums/{id}/tracks": {
            "get": {
                "description": "Returns all tracks for an album identified by a representative track ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "music"
                ],
                "summary": "List tracks for an album by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Album representative track ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_viewra_viewra_internal_application_music.ListTracksResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/music/artists": {
            "get": {
                "description": "Returns a list of all artists in a library with album and track counts and optional pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "music"
                ],
                "summary": "List music artists",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Library ID to filter artists",
                        "name": "library_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page (default: 50, max: 200)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to skip (default: 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_viewra_viewra_internal_application_music.ListArtistsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/music/artists/{id}/albums": {
            "get": {
                "description": "Returns all albums for an artist identified by a representative track ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "music"
                ],
                "summary": "List albums for an artist by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Artist representative track ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_viewra_viewra_internal_application_music.ListAlbumsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/music/search": {
            "get": {
                "description": "Searches for music tracks by title, artist, or album in a specific library",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "music"
                ],
                "summary": "Search music",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Library ID to search in",
                        "name": "library_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search query (title, artist, or album)",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_viewra_viewra_internal_application_music.ListTracksResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/music/tracks/{id}": {
            "get": {
                "description": "Returns details of a specific music track including all metadata",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "music"
                ],
                "summary": "Get a music track by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Track ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_viewra_viewra_internal_application_music.MusicTrackResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_viewra_viewra_internal_application_music.AlbumSummary": {
            "type": "object",
            "properties": {
                "album": {
                    "type": "string"
                },
                "artist": {
                    "type": "string"
                },
                "id": {
                    "description": "Representative media_id (first track)",
                    "type": "integer"
                },
                "track_count": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "github_com_viewra_viewra_internal_application_music.ArtistSummary": {
            "type": "object",
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "album_count": {
                    "type": "integer"
                },
                "id": {
                    "description": "Representative media_id (first track)",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "track_count": {
                    "type": "integer"
                }
            }
        },
        "github_com_viewra_viewra_internal_application_music.ListAlbumsResponse": {
            "type": "object",
            "properties": {
                "albums": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_viewra_viewra_internal_application_music.AlbumSummary"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_viewra_viewra_internal_domain_common.PaginationMetadata"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "github_com_viewra_viewra_internal_application_music.ListArtistsResponse": {
            "type": "object",
            "properties": {
                "artists": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_viewra_viewra_internal_application_music.ArtistSummary"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_viewra_viewra_internal_domain_common.PaginationMetadata"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "github_com_viewra_viewra_internal_application_music.ListTracksResponse": {
            "type": "object",
            "properties": {
                "pagination": {
                    "$ref": "#/definitions/github_com_viewra_viewra_internal_domain_common.PaginationMetadata"
                },
                "total": {
                    "type": "integer"
                },
                "tracks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_viewra_viewra_internal_application_music.MusicTrackResponse"
                    }
                }
            }
        },
        "github_com_viewra_viewra_internal_application_music.MusicTrackResponse": {
            "type": "object",
            "properties": {
                "album": {
                    "type": "string"
                },
                "album_artist": {
                    "type": "string"
                },
                "artist": {
                    "description": "Music Track-specific fields",
                    "type": "string"
                },
                "audio_codec": {
                    "type": "string"
                },
                "bitrate": {
                    "description": "Music-specific bitrate in kbps",
                    "type": "integer"
                },
                "composer": {
                    "type": "string"
                },
                "container_format": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "disc_number": {
                    "type": "integer"
                },
                "duration": {
                    "description": "in seconds",
                    "type": "integer"
                },
                "file_path": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "frame_rate": {
                    "type": "number"
                },
                "genre": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "description": "Base Media fields",
                    "type": "integer"
                },
                "is_extra": {
                    "type": "boolean"
                },
                "library_id": {
                    "type": "integer"
                },
                "media_bitrate": {
                    "description": "From base media",
                    "type": "integer"
                },
                "publisher": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "track_number": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "video_codec": {
                    "type": "string"
                },
                "width": {
                    "description": "Technical metadata",
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_viewra_viewra_internal_domain_common.PaginationMetadata": {
            "type": "object",
            "properties": {
                "hasMore": {
                    "description": "Whether there are more items available",
                    "type": "boolean"
                },
                "limit": {
                    "description": "Number of items per page",
                    "type": "integer"
                },
                "offset": {
                    "description": "Number of items skipped",
                    "type": "integer"
                },
                "total": {
                    "description": "Total number of items across all pages",
                    "type": "integer"
                }
            }
        },
...
        "internal_api_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}